Move WorkingTimeSettings into the worktingtime package
