+++ src/main/java/org/synyx/urlaubsverwaltung/workingtime/FederalState.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.settings;
+package org.synyx.urlaubsverwaltung.workingtime;
 
 import java.util.Arrays;
