+++ src/test/java/org/synyx/urlaubsverwaltung/settings/SettingsViewControllerTest.java
@@ -11,6 +11,7 @@ import org.springframework.validation.Errors;
 import org.synyx.urlaubsverwaltung.absence.Absence;
 import org.synyx.urlaubsverwaltung.calendarintegration.providers.CalendarProvider;
 import org.synyx.urlaubsverwaltung.period.DayLength;
+import org.synyx.urlaubsverwaltung.workingtime.FederalState;
 
 import java.time.Clock;
 import java.util.List;
