+++ src/test/java/org/synyx/urlaubsverwaltung/workingtime/WorkingTimeSettingsTest.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.settings;
+package org.synyx.urlaubsverwaltung.workingtime;
 
 import org.junit.jupiter.api.Test;
 import org.synyx.urlaubsverwaltung.period.DayLength;
